requirements.txt
SciPy
matplotlib.pyplot
numpy
pandas